The board is now scrollable.
To prevent an AssertionError from Pyglet, I had to check, before deleting a VertexList, that its attribute domain.allocator.starts evaluates to True. domain.allocator.starts is actually a List--I think it's a list of indices into the VertexList?--and if it's empty (therefore, False), I think that means you have a vertex list with nothing to draw, and therefore, nothing to delete.

+++ gui.py
@@ -9,10 +9,18 @@ class GameWindow:
     arrowhead_angle = 45
     arrowhead_len = 10
 
+    def __getattr__(self, attrn):
+        if attrn == 'width':
+            return self.window.width
+        elif attrn == 'height':
+            return self.window.height
+        else:
+            raise AttributeError(
+                "GameWindow has no attribute named {0}".format(attrn))
+
     def __init__(self, gamestate, boardname, batch=None):
         self.db = gamestate.db
         self.gamestate = gamestate
-        self.board = self.db.boarddict[boardname]
 
         self.boardgroup = pyglet.graphics.OrderedGroup(0)
         self.edgegroup = pyglet.graphics.OrderedGroup(1)
@@ -33,37 +41,25 @@ class GameWindow:
         self.mouse_dy = 0
         self.mouse_buttons = 0
         self.mouse_mods = 0
-        self.view_left = 0
         self.prev_view_left = 0
         self.prev_view_bot = 0
-        self.view_bot = 0
-
-        self.to_mouse = (
-            self.board.pawndict.values() +
-            self.board.spotdict.values())
-        for menu in self.board.menudict.itervalues():
-            self.to_mouse.extend(menu.items)
 
         window = pyglet.window.Window()
         if batch is None:
             batch = pyglet.graphics.Batch()
 
         self.window = window
-        self.width = self.window.width
-        self.height = self.window.height
         self.batch = batch
-        self.menus = self.board.menudict.values()
-        self.spots = self.board.spotdict.values()
-        self.pawns = self.board.pawndict.values()
+
+        self.board = self.db.boarddict[boardname]
+        self.board.set_gw(self)
+
         self.calcols = []
-        for pawn in self.pawns:
+        for pawn in self.board.pawndict.itervalues():
             if hasattr(pawn, 'calcol'):
                 self.calcols.append(pawn.calcol)
 
         self.calendar = self.board.calendar
-        self.calendar.set_gw(self)
-        for menu in self.board.menudict.itervalues():
-            menu.set_gw(self)
         self.drawn_board = None
         self.drawn_edges = None
         self.timeline = None
@@ -77,39 +73,35 @@ class GameWindow:
             """Draw the background image; all spots, pawns, and edges on the
 board; all visible menus; and the calendar, if it's visible."""
             # draw the background image
-            x = -1 * self.view_left
-            y = -1 * self.view_bot
             s = pyglet.sprite.Sprite(
-                self.board.wallpaper, x, y,
+                self.board.wallpaper,
+                self.board.offset_x,
+                self.board.offset_y,
                 batch=self.batch, group=self.boardgroup)
-            try:
-                self.drawn_board.delete()
-            except AttributeError:
-                pass
             self.drawn_board = s
-            # if the board has moved since last frame, then
-            # so has everything on it.
-            # redraw_all = (self.view_left != self.prev_view_left or
-            #               self.view_bot != self.prev_view_bot)
-            # if redraw_all:
-            #     self.onscreen = set()
-            # self.prev_view_left = self.view_left
-            # self.prev_view_bot = self.view_bot
             # draw the edges, representing portals
             e = []
             for dests in self.board.dimension.portalorigdestdict.itervalues():
                 for port in dests.itervalues():
                     origspot = port.orig.spot
                     destspot = port.dest.spot
-                    edge = (origspot.x, origspot.y, destspot.x, destspot.y)
-                    e.extend(edge)
-            try:
-                self.drawn_edges.delete()
-            except AttributeError:
-                pass
-            self.drawn_edges = self.batch.add(
-                len(e) / 2, pyglet.graphics.GL_LINES,
-                self.edgegroup, ('v2i', e))
+                    if origspot.visible or destspot.visible:
+                        edge = (origspot.window_x,
+                                origspot.window_y,
+                                destspot.window_x,
+                                destspot.window_y)
+                        e.extend(edge)
+            if (
+                    self.drawn_edges is not None and
+                    self.drawn_edges.domain.allocator.starts):
+                try:
+                    self.drawn_edges.delete()
+                except AttributeError:
+                    pass
+            if len(e) > 0:
+                self.drawn_edges = self.batch.add(
+                    len(e) / 2, pyglet.graphics.GL_LINES,
+                    self.edgegroup, ('v2i', e))
             # draw the spots, representing places
             for spot in self.board.spotdict.itervalues():
                 newstate = spot.get_state_tup()
@@ -118,13 +110,17 @@ board; all visible menus; and the calendar, if it's visible."""
                 self.onscreen.discard(spot.oldstate)
                 self.onscreen.add(newstate)
                 spot.oldstate = newstate
-                try:
-                    spot.sprite.delete()
-                except AttributeError:
-                    pass
+                if spot.sprite is not None:
+                    try:
+                        spot.sprite.delete()
+                    except AttributeError:
+                        pass
                 if spot.visible:
                     spot.sprite = pyglet.sprite.Sprite(
-                        spot.img, spot.left, spot.bot, batch=self.batch,
+                        spot.img,
+                        spot.window_left,
+                        spot.window_bot,
+                        batch=self.batch,
                         group=self.spotgroup)
             # draw the pawns, representing things
             for pawn in self.board.pawndict.itervalues():
@@ -134,10 +130,11 @@ board; all visible menus; and the calendar, if it's visible."""
                 self.onscreen.discard(pawn.oldstate)
                 self.onscreen.add(newstate)
                 pawn.oldstate = newstate
-                try:
-                    pawn.sprite.delete()
-                except AttributeError:
-                    pass
+                if pawn.sprite is not None:
+                    try:
+                        pawn.sprite.delete()
+                    except AttributeError:
+                        pass
                 if pawn.visible:
                     pawn.sprite = pyglet.sprite.Sprite(
                         pawn.img, pawn.left, pawn.bot, batch=self.batch,
@@ -151,10 +148,11 @@ board; all visible menus; and the calendar, if it's visible."""
                     self.onscreen.discard(menu_item.oldstate)
                     self.onscreen.add(newstate)
                     menu_item.oldstate = newstate
-                    try:
-                        menu_item.label.delete()
-                    except AttributeError:
-                        pass
+                    if menu_item.label is not None:
+                        try:
+                            menu_item.label.delete()
+                        except AttributeError:
+                            pass
                     if menu_item.visible:
                         sty = menu.style
                         if menu_item.hovered:
@@ -176,10 +174,11 @@ board; all visible menus; and the calendar, if it's visible."""
                 self.onscreen.discard(menu.oldstate)
                 self.onscreen.add(newstate)
                 menu.oldstate = newstate
-                try:
-                    menu.sprite.delete()
-                except AttributeError:
-                    pass
+                if menu.sprite is not None:
+                    try:
+                        menu.sprite.delete()
+                    except AttributeError:
+                        pass
                 if menu.visible:
                     image = (
                         menu.inactive_pattern.create_image(
@@ -196,10 +195,11 @@ board; all visible menus; and the calendar, if it's visible."""
                 self.onscreen.discard(self.calendar.oldstate)
                 self.calendar.oldstate = newstate
                 for calcol in self.calcols:
-                    try:
-                        calcol.sprite.delete()
-                    except AttributeError:
-                        pass
+                    if calcol.sprite is not None:
+                        try:
+                            calcol.sprite.delete()
+                        except AttributeError:
+                            pass
                     if calcol.visible:
                         image = calcol.inactive_pattern.create_image(
                             calcol.width, calcol.height)
@@ -207,14 +207,16 @@ board; all visible menus; and the calendar, if it's visible."""
                             image, calcol.left, calcol.bot,
                             batch=self.batch, group=self.calendargroup)
                     for cel in calcol.celldict.itervalues():
-                        try:
-                            cel.sprite.delete()
-                        except AttributeError:
-                            pass
-                        try:
-                            cel.label.delete()
-                        except AttributeError:
-                            pass
+                        if cel.sprite is not None:
+                            try:
+                                cel.sprite.delete()
+                            except AttributeError:
+                                pass
+                        if cel.label is not None:
+                            try:
+                                cel.label.delete()
+                            except AttributeError:
+                                pass
                         if cel.visible:
                             if self.hovered == cel:
                                 pat = cel.active_pattern
@@ -235,7 +237,9 @@ board; all visible menus; and the calendar, if it's visible."""
                                 batch=self.batch, group=self.labelgroup)
             if self.last_age != self.gamestate.age:
                 # draw the time line on top of the calendar
-                if self.last_timeline_y > 0:
+                if (
+                        self.timeline is not None and
+                        self.timeline.domain.allocator.starts):
                     try:
                         self.timeline.delete()
                     except AttributeError:
@@ -264,8 +268,9 @@ board; all visible menus; and the calendar, if it's visible."""
             """Find the widget, if any, that the mouse is over, and highlight
 it."""
             if self.hovered is None:
-                for menu in self.menus:
+                for menu in self.board.menudict.itervalues():
                     if (
+                            menu.visible and
                             x > menu.left and
                             x < menu.right and
                             y > menu.bot and
@@ -280,18 +285,20 @@ it."""
                                     item.set_hovered()
                                 self.hovered = item
                                 return
-                for spot in self.spots:
+                for spot in self.board.spotdict.itervalues():
                     if (
-                            x > spot.left and
-                            x < spot.right and
-                            y > spot.bot and
-                            y < spot.top):
+                            spot.visible and
+                            x > spot.window_left and
+                            x < spot.window_right and
+                            y > spot.window_bot and
+                            y < spot.window_top):
                         if hasattr(spot, 'set_hovered'):
                             spot.set_hovered()
                         self.hovered = spot
                         return
-                for pawn in self.pawns:
+                for pawn in self.board.pawndict.itervalues():
                     if (
+                            pawn.visible and
                             x > pawn.left and
                             x < pawn.right and
                             y > pawn.bot and
@@ -325,20 +332,22 @@ still over it when pressed, it's been half-way clicked; remember this."""
         def on_mouse_release(x, y, button, modifiers):
             """If something was being dragged, drop it. If something was being
 pressed but not dragged, it's been clicked. Otherwise do nothing."""
-            if self.grabbed is not None:
+            if self.grabbed is None:
+                self.board.save()
+            else:
                 if hasattr(self.grabbed, 'dropped'):
                     self.grabbed.dropped(x, y, button, modifiers)
                 self.grabbed = None
-            elif (self.pressed is not None and
-                  x > self.pressed.left and
-                  x < self.pressed.right and
-                  y > self.pressed.bot and
-                  y < self.pressed.top and
-                  hasattr(self.pressed, 'onclick')):
-                self.pressed.onclick(button, modifiers)
             if self.pressed is not None:
                 if hasattr(self.pressed, 'unset_pressed'):
                     self.pressed.unset_pressed()
+                if (
+                        hasattr(self.pressed, 'onclick') and
+                        x > self.pressed.left and
+                        x < self.pressed.right and
+                        y > self.pressed.bot and
+                        y < self.pressed.top):
+                    self.pressed.onclick()
                 self.pressed = None
 
         @window.event
@@ -347,20 +356,36 @@ pressed but not dragged, it's been clicked. Otherwise do nothing."""
 move_with_mouse method, use it.
 
             """
-            if self.grabbed is not None:
-                self.grabbed.move_with_mouse(x, y, dx, dy, buttons, modifiers)
-            elif (self.pressed is not None and
-                  x > self.pressed.left and
-                  x < self.pressed.right and
-                  y > self.pressed.bot and
-                  y < self.pressed.top and
-                  hasattr(self.pressed, 'move_with_mouse')):
-                self.grabbed = self.pressed
+            if self.grabbed is None:
+                if (
+                        self.pressed is not None and
+                        x > self.pressed.left and
+                        x < self.pressed.right and
+                        y > self.pressed.bot and
+                        y < self.pressed.top and
+                        hasattr(self.pressed, 'move_with_mouse')):
+                    self.grabbed = self.pressed
+                else:
+                    if self.board.view_left - dx < 0 or (
+                            (self.board.view_left - dx)
+                            + self.window.width > self.board.width):
+                        effective_dx = 0
+                    else:
+                        effective_dx = dx
+                    if self.board.view_bot - dy < 0 or (
+                            (self.board.view_bot - dy)
+                            + self.window.height > self.board.height):
+                        effective_dy = 0
+                    else:
+                        effective_dy = dy
+                    self.board.view_left -= effective_dx
+                    self.board.view_bot -= effective_dy
+                    if self.pressed is not None:
+                        self.pressed.unset_pressed()
+                        self.pressed = None
+                    self.grabbed = None
             else:
-                if self.pressed is not None:
-                    self.pressed.unset_pressed()
-                    self.pressed = None
-                self.grabbed = None
+                self.grabbed.move_with_mouse(x, y, dx, dy, buttons, modifiers)
 
         @window.event
         def on_resize(w, h):
@@ -382,4 +407,3 @@ move_with_mouse method, use it.
                 if self.calendar.scrolled_to < 0:
                     self.calendar.scrolled_to = 0
                 self.board.save()
-                print "calendar scrolled to %d" % self.calendar.scrolled_to
